# GTiff: fix speed of guessing JPEG quality