--- a/gdal/frmts/gtiff/GNUmakefile
+++ b/gdal/frmts/gtiff/GNUmakefile
@@ -69,3 +69,8 @@ lib-geotiff:
 	(cd libgeotiff; $(MAKE) install-obj)
 
 install-obj:	$(SUBLIBS) $(O_OBJ:.o=.$(OBJ_EXT))
+
+generate_quant_table_md5sum: generate_quant_table_md5sum.cpp
+	$(CXX) -O2 -g -Wall generate_quant_table_md5sum.cpp -o generate_quant_table_md5sum -I../../port -I../../gcore -I../../ogr -L../../ -lgdal
+	./generate_quant_table_md5sum > quant_table_md5sum.h
+	$(RM) -f generate_quant_table_md5sum