… gdal/frmts/gtiff/geotiff.cpp
… gdal/frmts/gtiff/quant_table_md5sum.h