@rouault 2021-04-27
Actually the related peformance regression started at the beginning of the 3.3 cycle, and did not affect 3.2